Split activity feed launch helpers
