+++ Sources/AppDelegate.swift
@@ -945,92 +945,6 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
         return true
     }
 
-    private static func workstreamPersistenceFileURLForLaunch(env: [String: String]) -> URL {
-#if DEBUG
-        if let path = env["CMUX_UI_TEST_WORKSTREAM_FILE"]?.trimmingCharacters(in: .whitespacesAndNewlines),
-           !path.isEmpty {
-            return URL(fileURLWithPath: path)
-        }
-#endif
-        return WorkstreamPersistence.defaultFileURL()
-    }
-
-    private static func workstreamInitialLoadLimitForLaunch(env: [String: String]) -> Int {
-#if DEBUG
-        if let raw = env["CMUX_UI_TEST_WORKSTREAM_INITIAL_LOAD_LIMIT"],
-           let value = Int(raw),
-           value > 0 {
-            return value
-        }
-#endif
-        return WorkstreamDefaultInitialLoadLimit
-    }
-
-    private static func workstreamHistoryPageSizeForLaunch(env: [String: String]) -> Int {
-#if DEBUG
-        if let raw = env["CMUX_UI_TEST_WORKSTREAM_HISTORY_PAGE_SIZE"],
-           let value = Int(raw),
-           value > 0 {
-            return value
-        }
-#endif
-        return WorkstreamDefaultHistoryPageSize
-    }
-
-#if DEBUG
-    private static func seedFeedActivityPaginationUITestLogIfNeeded(
-        env: [String: String],
-        fileURL: URL
-    ) {
-        guard let rawCount = env["CMUX_UI_TEST_FEED_ACTIVITY_SEED_COUNT"],
-              let count = Int(rawCount),
-              count > 0 else {
-            return
-        }
-
-        let encoder = JSONEncoder()
-        encoder.dateEncodingStrategy = .iso8601
-        let baseDate = Date(timeIntervalSince1970: 1_700_000_000)
-        var data = Data()
-        do {
-            for index in 0..<count {
-                let item = WorkstreamItem(
-                    workstreamId: "opencode-ui-page-\(index)",
-                    source: .opencode,
-                    kind: .todos,
-                    createdAt: baseDate.addingTimeInterval(TimeInterval(index)),
-                    cwd: "/tmp/cmux-feed-pagination",
-                    status: .telemetry,
-                    payload: .todos([]),
-                    context: WorkstreamContext(lastUserMessage: "activity pagination seed \(index)")
-                )
-                data.append(try encoder.encode(item))
-                data.append(0x0A)
-            }
-            try FileManager.default.createDirectory(
-                at: fileURL.deletingLastPathComponent(),
-                withIntermediateDirectories: true
-            )
-            try data.write(to: fileURL, options: .atomic)
-            _ = CmuxUITestCapture.mutateJSONObjectIfConfigured(
-                envKey: "CMUX_UI_TEST_FEED_ACTIVITY_PAGINATION_PATH"
-            ) { payload in
-                payload["seeded"] = "1"
-                payload["seedCount"] = "\(count)"
-                payload["seedPath"] = fileURL.path
-            }
-        } catch {
-            _ = CmuxUITestCapture.mutateJSONObjectIfConfigured(
-                envKey: "CMUX_UI_TEST_FEED_ACTIVITY_PAGINATION_PATH"
-            ) { payload in
-                payload["seeded"] = "0"
-                payload["seedError"] = "\(error)"
-                payload["seedPath"] = fileURL.path
-            }
-        }
-    }
-#endif
-
     func applicationDidFinishLaunching(_ notification: Notification) {
         let env = ProcessInfo.processInfo.environment
         let isRunningUnderXCTest = isRunningUnderXCTest(env)
@@ -1044,26 +958,11 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
 
         claimAuthCallbackURLSchemes()
 
-        let workstreamPersistenceURL = Self.workstreamPersistenceFileURLForLaunch(env: env)
-#if DEBUG
-        Self.seedFeedActivityPaginationUITestLogIfNeeded(
-            env: env,
-            fileURL: workstreamPersistenceURL
-        )
-#endif
-
         // Install the Feed (workstream) store. Separate from the transport
         // wiring: the store is a plain singleton here, and the socket
         // `feed.*` V2 verbs in `TerminalController` push into it directly
         // via `FeedCoordinator`.
-        FeedCoordinator.shared.install(
-            store: WorkstreamStore(
-                transport: NullWorkstreamTransport(),
-                persistence: WorkstreamPersistence(fileURL: workstreamPersistenceURL),
-                initialLoadLimit: Self.workstreamInitialLoadLimitForLaunch(env: env),
-                historyPageSize: Self.workstreamHistoryPageSizeForLaunch(env: env)
-            )
-        )
+        Self.installWorkstreamStoreForLaunch(env: env)
         Task { @MainActor in
             await FeedCoordinator.shared.store?.start()
 #if DEBUG
