+++ Sources/Feed/FeedDisplayLabels.swift
@@ -1,0 +1,47 @@
+import CMUXWorkstream
+import Foundation
+
+extension WorkstreamPermissionMode {
+    var displayLabel: String {
+        switch self {
+        case .once:
+            return String(localized: "feed.permission.mode.once", defaultValue: "once")
+        case .always:
+            return String(localized: "feed.permission.mode.always", defaultValue: "always")
+        case .all:
+            return String(localized: "feed.permission.mode.all", defaultValue: "all tools")
+        case .bypass:
+            return String(localized: "feed.permission.mode.bypass", defaultValue: "bypass")
+        case .deny:
+            return String(localized: "feed.permission.mode.deny", defaultValue: "denied")
+        }
+    }
+}
+
+extension WorkstreamExitPlanMode {
+    var displayLabel: String {
+        switch self {
+        case .ultraplan:
+            return String(localized: "feed.exitplan.mode.ultraplan", defaultValue: "ultraplan")
+        case .bypassPermissions:
+            return String(localized: "feed.exitplan.mode.bypass", defaultValue: "bypass")
+        case .autoAccept:
+            return String(localized: "feed.exitplan.mode.autoAccept", defaultValue: "auto")
+        case .manual:
+            return String(localized: "feed.exitplan.mode.manual", defaultValue: "manual")
+        case .deny:
+            return String(localized: "feed.exitplan.mode.deny", defaultValue: "denied")
+        }
+    }
+}
+
+extension WorkstreamSource {
+    var feedDisplayName: String {
+        switch self {
+        case .opencode:
+            return String(localized: "sessionIndex.agent.opencode", defaultValue: "OpenCode")
+        default:
+            return rawValue.capitalized
+        }
+    }
+}
