+++ Sources/Feed/FeedPanelView.swift
@@ -10,51 +10,6 @@ private func feedDebugResponderSummary(_ responder: NSResponder?) -> String {
 }
 #endif
 
-private extension WorkstreamPermissionMode {
-    var displayLabel: String {
-        switch self {
-        case .once:
-            return String(localized: "feed.permission.mode.once", defaultValue: "once")
-        case .always:
-            return String(localized: "feed.permission.mode.always", defaultValue: "always")
-        case .all:
-            return String(localized: "feed.permission.mode.all", defaultValue: "all tools")
-        case .bypass:
-            return String(localized: "feed.permission.mode.bypass", defaultValue: "bypass")
-        case .deny:
-            return String(localized: "feed.permission.mode.deny", defaultValue: "denied")
-        }
-    }
-}
-
-private extension WorkstreamExitPlanMode {
-    var displayLabel: String {
-        switch self {
-        case .ultraplan:
-            return String(localized: "feed.exitplan.mode.ultraplan", defaultValue: "ultraplan")
-        case .bypassPermissions:
-            return String(localized: "feed.exitplan.mode.bypass", defaultValue: "bypass")
-        case .autoAccept:
-            return String(localized: "feed.exitplan.mode.autoAccept", defaultValue: "auto")
-        case .manual:
-            return String(localized: "feed.exitplan.mode.manual", defaultValue: "manual")
-        case .deny:
-            return String(localized: "feed.exitplan.mode.deny", defaultValue: "denied")
-        }
-    }
-}
-
-private extension WorkstreamSource {
-    var feedDisplayName: String {
-        switch self {
-        case .opencode:
-            return String(localized: "sessionIndex.agent.opencode", defaultValue: "OpenCode")
-        default:
-            return rawValue.capitalized
-        }
-    }
-}
-
 /// Right-sidebar Feed view. Matches the Sessions page visual language:
 /// compact rows with SF Symbol + 13pt title + secondary metadata,
 /// full-width hover backgrounds, and control-bar pill buttons styled
@@ -152,33 +107,6 @@ struct FeedPanelView: View {
         }
 #endif
         return .actionable
-    }
-}
-
-private struct FeedSecondaryFilterButton: View {
-    let filter: FeedPanelView.Filter
-    let isSelected: Bool
-    let action: () -> Void
-    @State private var isHovered = false
-
-    var body: some View {
-        Button(action: action) {
-            HStack(spacing: 3) {
-                Image(systemName: filter.symbolName)
-                    .font(.system(size: 10, weight: .medium))
-                Text(filter.label)
-                    .font(.system(size: 11, weight: .medium))
-            }
-            .rightSidebarChromePill(
-                isSelected: isSelected,
-                isHovered: isHovered,
-                geometryKeyPrefix: "rightSidebarSecondaryControl_feed_\(filter.rawValue)"
-            )
-        }
-        .buttonStyle(.plain)
-        .onHover { isHovered = $0 }
-        .help(filter.label)
-        .accessibilityIdentifier("FeedFilterButton.\(filter.rawValue)")
     }
 }
