+++ Sources/Feed/FeedSecondaryFilterButton.swift
@@ -1,0 +1,28 @@
+import SwiftUI
+
+struct FeedSecondaryFilterButton: View {
+    let filter: FeedPanelView.Filter
+    let isSelected: Bool
+    let action: () -> Void
+    @State private var isHovered = false
+
+    var body: some View {
+        Button(action: action) {
+            HStack(spacing: 3) {
+                Image(systemName: filter.symbolName)
+                    .font(.system(size: 10, weight: .medium))
+                Text(filter.label)
+                    .font(.system(size: 11, weight: .medium))
+            }
+            .rightSidebarChromePill(
+                isSelected: isSelected,
+                isHovered: isHovered,
+                geometryKeyPrefix: "rightSidebarSecondaryControl_feed_\(filter.rawValue)"
+            )
+        }
+        .buttonStyle(.plain)
+        .onHover { isHovered = $0 }
+        .help(filter.label)
+        .accessibilityIdentifier("FeedFilterButton.\(filter.rawValue)")
+    }
+}
